01-inventory-sqlbot/requirements.txt
streamlit
python-dotenv
sqlalchemy
pymysql
langchain
langchain-community
langchain-core
langchain-ollama
chromadb
sentence-transformers
